MF-967 - Move events stuff to subdir
