+++ things/redis/events/events.go
@@ -1,4 +1,4 @@
-package redis
+package events
 
 import (
 	"encoding/json"
