+++ things/redis/events/streams.go
@@ -1,7 +1,7 @@
 // Copyright (c) Mainflux
 // SPDX-License-Identifier: Apache-2.0
 
-package redis
+package events
 
 import (
 	"context"
